Write in file

+++ api/src/main/java/ar/edu/itba/pod/query3/UnemploymentIndexCollator.java
@@ -2,23 +2,18 @@ package ar.edu.itba.pod.query3;
 
 import com.hazelcast.mapreduce.Collator;
 
-import java.math.BigDecimal;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
-public class UnemploymentIndexCollator implements Collator<Map.Entry<String, BigDecimal>, Map<String, BigDecimal>> {
+public class UnemploymentIndexCollator implements Collator<Map.Entry<String, Double>, Map<String, Double>>{
     @Override
-    public Map<String, BigDecimal> collate(Iterable<Map.Entry<String, BigDecimal>> iterable) {
-        List<Map.Entry<String, BigDecimal>> list = StreamSupport.stream(iterable.spliterator(), false)
+    public Map<String, Double> collate(Iterable<Map.Entry<String, Double>> iterable) {
+        List<Map.Entry<String, Double>> list = StreamSupport.stream(iterable.spliterator(), false)
                 .collect(Collectors.toList());
-        //Collections.sort(list, Comparator.comparing(Map.Entry::getValue));
         Collections.sort(list, (o1, o2) -> o2.getValue().compareTo(o1.getValue()));
-        Map<String, BigDecimal> result = new LinkedHashMap<>();
-        for (Map.Entry<String, BigDecimal> entry: list) {
+        Map<String, Double> result = new LinkedHashMap<>();
+        for (Map.Entry<String, Double> entry : list){
             result.put(entry.getKey(), entry.getValue());
         }
         return result;
